refactors a bunch of the testing suite and moves tests into a separate folder, removes xml files'
