--- .idea/libraries/dnaAnalyzer_node_modules.xml
@@ -1,14 +1,0 @@
-<component name="libraryTable">
-  <library name="dnaAnalyzer node_modules" type="javaScript">
-    <properties>
-      <option name="frameworkName" value="node_modules" />
-      <sourceFilesUrls>
-        <item url="file://$PROJECT_DIR$/node_modules" />
-      </sourceFilesUrls>
-    </properties>
-    <CLASSES>
-      <root url="file://$PROJECT_DIR$/node_modules" />
-    </CLASSES>
-    <SOURCES />
-  </library>
-</component>
