+++ browser/js/analyzer/analyzer.html
@@ -1,11 +1,8 @@
 <div class="container clarity">
 	<div class="text-center">
-		<h2> Test Results</h2>
+		<button ui-sref="home" class='back-button btn'>Back to Dashboard</button>
+		<h2 class='header'> Test Results</h2>
 	</div>
-	<!-- <div class="col-md-12">
-			<img class="chromo" src="../../../public/chromosomes.jpg">
-	</div> -->
-	
 	<hr>
 	<div class="text-center">
 		<h4> Overall Health Score: {{total}} </h4>
