+++ browser/js/home/home.html
@@ -24,7 +24,7 @@
     <div ng-show="converting" class='landing'>
 		<div class="col-md-12">
             <h3 class="clarity">Uploading File ...</h3>
-            <h5 class="clarity">* this can take up to 20min - thanks for you patience *</h5>
+            <h5 class="clarity">* this can take up to 45min - thanks for your patience *</h5>
         </div>
     <div>
 </div>
